Add files via upload
Hi there,
This project is a demonstration of Inheritance using a real life example. First i've developed a basic car structure with basic minimum requirements for a functional car. This will act as a blueprint for the final vehicle of ours which is to be developed. Having inherited the basic car elements from the car.java file, i've added some extra elements which are specific to the final car which is built.

Applicatons:
This analogy can be used as a blueprint in industries to build a complex vehicle right from the scratch. 
Since the basic elements are always the same, the code is reused everytime we build a program for any kind of the vehicle (in our case car).
Suggestions are most welcome
Regards, 
Ravi Anand